fix after deployment add condition for local or production
