+++ .env
@@ -1,0 +1,3 @@
+NEXT_PUBLIC_PROJECT_NAME= pokemon
+NEXT_PUBLIC_ENV= local
+#NEXT_PUBLIC_ENV= production
