+++ components/templates/layout.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import Head from "next/head";
 import NavigationHeader from "Components/molecules/navigation-header";
+import { getUrl } from "Helpers/common-helper";
 import ViewSwitch from "Components/templates/viewswitch";
 
 interface ILayout {
@@ -12,7 +13,7 @@ const Layout = (props: ILayout) => {
     <div className="relative">
       <Head>
         <title>{props.title}</title>
-        <link rel="icon" href="/pokeball.png" />
+        <link rel="icon" href={getUrl("/pokeball.png")} />
       </Head>
       <NavigationHeader />
       <ViewSwitch
