+++ pages/404.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import Layout from "Components/templates/layout";
+
+const PokemonList = () => {
+  return (
+    <Layout title="Pokemon List">
+      <div className="border my-10 p-10 sm:p-40 flex justify-center font-bold sm:text-2xl text-red-500 rounded">
+        PAGE NOT FOUND
+      </div>
+    </Layout>
+  );
+};
+
+export default PokemonList;
